requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlmodel>=0.0.8
jinja2>=3.1.2
python-multipart>=0.0.5
pandas>=1.3.0
openpyxl>=3.0.9
reportlab>=3.6.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.0